fix(code-review): automated fixer cycle — PR #2013

+++ supabase/migrations/20260829000001_metering_ask_columns.sql
@@ -7,11 +7,16 @@
 -- ZEROS (never 500, P2-14). The increment RPC mirrors metering_increment
 -- (0014): atomic under Postgres row locking, SECURITY DEFINER,
 -- service_role-only.
+--
+-- ask_tokens_in/out are bigint (NOT integer, P2): the ask envelope is up to
+-- ~8k tokens/query x 60/min/team sustained → ~20.7B tokens/month, ~10x over
+-- the integer range — overflow would raise and (best-effort metering)
+-- silently stop recording.
 
 ALTER TABLE public.metering_records
     ADD COLUMN IF NOT EXISTS ask_calls       integer NOT NULL DEFAULT 0,
-    ADD COLUMN IF NOT EXISTS ask_tokens_in   integer NOT NULL DEFAULT 0,
-    ADD COLUMN IF NOT EXISTS ask_tokens_out  integer NOT NULL DEFAULT 0,
+    ADD COLUMN IF NOT EXISTS ask_tokens_in   bigint  NOT NULL DEFAULT 0,
+    ADD COLUMN IF NOT EXISTS ask_tokens_out  bigint  NOT NULL DEFAULT 0,
     ADD COLUMN IF NOT EXISTS ask_cost_usd    double precision NOT NULL DEFAULT 0;
 
 CREATE OR REPLACE FUNCTION public.metering_increment_ask(
